+++ rstar/CHANGELOG.md
@@ -2,10 +2,22 @@
 
 ## Added
 - Added missing re-exports of nearest neighbor iterators so they can be named in downstream crates. ([PR](https://github.com/georust/rstar/pull/186))
+- Added nearest neighbor search with distance squared: `nearest_neighbor_with_distance_2` and `nearest_neighbors_with_distance_2` methods ([PR](https://github.com/georust/rstar/pull/191)).
+- Added `root` doc examples and traversal docs
+- Implemented `RTreeObject` for `Arc<T>` and `Rc<T>`
 - Added `Envelope::is_empty`. ([PR](https://github.com/georust/rstar/pull/190))
+- New `AABB::from_center` utility constructor
+
+## Fixed
+- Fix excessive memory retention in `bulk_load` from `Vec::split_off` over-capacity
+- Fix transcription error in implementation of OMT bulk loading leading to nodes which are too large.
 
 ## Changed
 - Made `RStar` methods take `Point` and `Envelope` as owned values where it makes sense ([PR](https://github.com/georust/rstar/pull/189))
+- Fix Clippy warning (surfaced in Rust 1.89) related to lifetime elision
+- MSRV is now 1.68 (2023-03-09)
+- Fix incorrect assertion message in `verify_parameters` of `rstar/src/params.rs`
+
 
 # 0.12.2